requirements.txt
dj-static==0.0.6
Django==2.2.17
django-stdimage==5.1.1
gunicorn==20.0.4
Pillow==8.1.0
progressbar2==3.53.1
python-utils==2.4.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
